requirements.txt
Flask
elasticsearch
python-dotenv
pytest
pytest-flask
gunicorn
pymongo
redis